\name{bendability}
\alias{bendability}
\title{bendability}
\description{spline function to calulate the profile of DNA 
propensity to bend for long sequences according to
I. Brukner, R. Sánchez, D. Suck, S. Pongor, The EMBO Journal 14, 1812–1818 (1995).
=======================================================}
\usage{bendability(s, bound, width = 1)}
\arguments{
  \item{s}{DNA sequence}
  \item{bound}{define a fragment of interest}
  \item{width}{smoothing window width}
}



\author{Anatoly Sorokin, Evgenia Temlyakova, Timur Dzhelyadin}





