\name{filesave}
\alias{filesave}
\title{Function for saving plots into a few file formats}
\description{It is used in bodies of several plot functions of the package. Possible formats 
are png, pdf, postscript, bmp, jpeg. There are some special settings for png to make 
good-for-our-wiki look of saved picture. }
\usage{filesave(imagefun, filetype, width = 1000, height = 1000)}
\arguments{
  \item{imagefun}{function, which output is a plot}
  \item{filetype}{character string, specifies file type for saving; possible formats are png, pdf, postscript, jpeg, bmp.}
  \item{width}{width of the plot}
  \item{height}{height of the plot}
}



\author{Anatoly Sorokin, Evgenia Temlyakova, Timur Dzhelyadin}





