\name{freeG}
\alias{freeG}
\title{freeG}
\description{function to calulate the profile of DNA free energy for long sequences
according to A. Kanhere, M. Bansal, BMC Bioinformatics 6, 1 (2005).
=======================================================}
\usage{freeG(s, bound, width = 1)}
\arguments{
  \item{s}{DNA sequence}
  \item{bound}{define a fragment of interest}
  \item{width}{smoothing window width}
}



\author{Anatoly Sorokin, Evgenia Temlyakova, Timur Dzhelyadin}





