\name{getOligos}
\alias{getOligos}
\title{getOligos}
\description{Function to calculate geometry of the DNA double helix.   }
\usage{getOligos(s, len = 2)}
\arguments{
  \item{s}{sequence of the DNA span to be build}
  \item{len}{length of the oligonucleotide}
}



\author{Anatoly Sorokin, Evgenia Temlyakova, Timur Dzhelyadin}





