\name{lseqcurlib}
\alias{lseqcurlib}
\title{lseqcurlib}
\description{main function to calculate electrostatic profile of DNA}
\usage{lseqcurlib(s, bound, width, ref, filename = NA)}
\arguments{
  \item{s}{DNA sequence}
  \item{bound}{define a fragment of interest}
  \item{width}{smoothing window width}
  \item{ref}{reference position}
  \item{filename}{
}
}

\value{list containing a few electrostatic data: pot, mpot, risef, i1}

\author{Anatoly Sorokin, Evgenia Temlyakova, Timur Dzhelyadin}





