\name{makeMultivalueCyllinder}
\alias{makeMultivalueCyllinder}
\title{makeMultivalueCyllinder}
\description{Function to create a list of points on the
cyllinder around DNA helix in the positions identical to 
generated by our Coulomb calculations. File ofile could be used
with APBS data as follow:
multivalue test.csv result.dx test.surf.out
and test.surf.out could be analysed with makePhi2D function.
all data for calculation could be get from APBS in file}
\usage{makeMultivalueCyllinder(xC = 0, yC = 0, zC = 0, zLen, radius = 15, 
    ofile)}
\arguments{
  \item{xC}{Position of the  cyllinder center (first number in fgcent)}
  \item{yC}{Position of the cyllinder center (second number in fgcent)}
  \item{zC}{Position of the cyllinder center (last number in fgcent)}
  \item{zLen}{Length of the cyllinder (last number in fglen)}
  \item{radius}{Radius of the cyllinder}
  \item{ofile}{path to file to write}
}



\author{Anatoly Sorokin, Evgenia Temlyakova, Timur Dzhelyadin}





