\name{makePhi2D}
\alias{makePhi2D}
\title{makePhi2D}
\description{Function to convert results of APBS multivalue function into
2D matrix for further analysis.}
\usage{makePhi2D(file)}
\arguments{
  \item{file}{input file, result of multivalue invocation.}
}

\value{matrix with angle in rows and Z in columns}

\author{Anatoly Sorokin, Evgenia Temlyakova, Timur Dzhelyadin}





