\name{plotPotential2D}
\alias{plotPotential2D}
\title{plotPotential2D}

\usage{plotPotential2D(x, y, p, nbreaks = 100, palette = colorRampPalette(c("black", 
    "red", "yellow"), space = "rgb"))}
\arguments{
  \item{x}{
}
  \item{y}{
}
  \item{p}{
}
  \item{nbreaks}{
}
  \item{palette}{
}
}



\author{Anatoly Sorokin, Evgenia Temlyakova, Timur Dzhelyadin}





