\name{sseqspline1D.BP}
\alias{sseqspline1D.BP}
\title{sseqspline1D BP}
\description{spline function to calulate the profile of electrostatics for short sequences in base pairs}
\usage{sseqspline1D.BP(s, ref, zout = -540:179)}
\arguments{
  \item{s}{DNA sequence}
  \item{ref}{reference position}
  \item{zout}{exact coordinates in which values of the potential will be calculated.}
}

\value{potential profile values at the centre of the base pair}

\author{Anatoly Sorokin, Evgenia Temlyakova, Timur Dzhelyadin}





