\name{updateX3DNAparam}
\alias{updateX3DNAparam}
\title{updateX3DNAparam}
\description{Function to read ideal B-form parameters generated by X3DNA package
and update it according to our internal DNA geometry settings.   }
\usage{updateX3DNAparam(ifile, ofile = ifile)}
\arguments{
  \item{ifile}{path to file to read}
  \item{ofile}{path to file to write}
}



\author{Anatoly Sorokin, Evgenia Temlyakova, Timur Dzhelyadin}





