\name{dnaGeom}
\alias{dnaGeom}
\title{dnaGeom}
\description{Function to calculate geometry of the DNA double helix.   }
\usage{dnaGeom(s)}
\arguments{
  \item{s}{sequence of the DNA span to be build}
}

\value{object of class 'DNAgeom' with two slots: rise, which gives Z coordinate of the corresponding base pair, and 
twist, which gives orientation of the X axis of the base pair.}
\references{http://chem.rutgers.edu/~xiangjun/3DNA/images/bp_step_hel.gif}
\author{Anatoly Sorokin, Evgenia Temlyakova, Timur Dzhelyadin}





