\name{lseqcurlib1D}
\alias{lseqcurlib1D}
\title{lseqcurlib1D}
\description{main function to calculate electrostatic profile of DNA}
\usage{lseqcurlib1D(s, bound, width, ref, name)}
\arguments{
  \item{s}{DNA sequence}
  \item{bound}{define a fragment of interest}
  \item{width}{smoothing window width}
  \item{ref}{reference position}
  \item{name}{name of the library}
}

\value{list containing a few electrostatic data: pot, mpot, risef, i1}

\author{Anatoly Sorokin, Evgenia Temlyakova, Timur Dzhelyadin}





