\name{lseqspline1D}
\alias{lseqspline1D}
\title{lseqspline1D}
\description{spline function to calulate the profile of electrostatics for long sequences
!!!!! DO NOT USE. FUNCTION IS UNDER DEVELOPMENT !!!!!!!
=======================================================}
\usage{lseqspline1D(s, bound, width = 1, ref)}
\arguments{
  \item{s}{DNA sequence}
  \item{bound}{define a fragment of interest}
  \item{width}{smoothing window width}
  \item{ref}{reference position}
}



\author{Anatoly Sorokin, Evgenia Temlyakova, Timur Dzhelyadin}





