\name{makeMultivalueCyllinder}
\alias{makeMultivalueCyllinder}
\title{makeMultivalueCyllinder}
\description{Function to create a list of points on the
cyllinder around DNA helix in the positions identical to 
generated by our Coulomb calculations. File ofile could be used
with APBS data as follow:
multivalue test.csv result.dx test.surf.out
and test.surf.out could be analysed with makePhi2D function.}
\usage{makeMultivalueCyllinder(zMin, zMax, radius = 15, ofile)}
\arguments{
  \item{zMin}{Position of the first circule of cyllinder}
  \item{zMax}{Position of the last circule of cyllinder}
  \item{radius}{Radius of the cyllinder}
  \item{ofile}{path to file to write}
}



\author{Anatoly Sorokin, Evgenia Temlyakova, Timur Dzhelyadin}





