\name{reldna-package}
\alias{reldna-package}
\docType{package}
\title{R Electrostatics of DNA}
\description{Package to calculate 1D electrostatic profile of arbitrary DNA sequence}
\details{
\tabular{ll}{Package: \tab reldna\cr
Type: \tab Package\cr
Title: \tab R Electrostatics of DNA\cr
Version: \tab 0.0-3\cr
Date: \tab 2012-11-14\cr
Author: \tab Anatoly Sorokin, Evgenia Temlyakova, Timur Dzhelyadin\cr
Maintainer: \tab Evgenia Temlyakova <evgenia.teml@gmail.com>\cr
License: \tab GPL (>= 2)\cr
Copyright: \tab Anatoly Sorokin, Evgenia Temlyakova, Timur Dzhelyadin 2012\cr
Depends: \tab seqinr\cr
Suggests: \tab lattice, xtable\cr
Imports: \tab car\cr
LazyData: \tab true\cr}
}
\author{Anatoly Sorokin, Evgenia Temlyakova, Timur Dzhelyadin}

\keyword{ package }


