\name{sseqspline1D}
\alias{sseqspline1D}
\title{sseqspline1D}
\description{spline function to calulate the profile of electrostatics for short sequences}
\usage{sseqspline1D(s, ref, zout = -540:179)}
\arguments{
  \item{s}{DNA sequence}
  \item{ref}{reference position}
  \item{zout}{exact coordinates in which values of the potential will be calculated.}
}



\author{Anatoly Sorokin, Evgenia Temlyakova, Timur Dzhelyadin}





